Separate libserialize module
